[Write API]: Include event category and language in the JSON request.
Category is mandatory while Language is optional (the default en-US will
be used instead).
